Pridėjimo mygtukai, css fixes

+++ resources/views/AdminViewGroups.blade.php
@@ -3,22 +3,22 @@
     <div class="grid mt-6 grid-cols-1">
         @include('layouts.alert')
         @foreach($groupActivities as $groupActivity)
-            <h1 class="font-bold text-2xl mt-6">Grupinė veikla "{{$groupActivity->title}}"</h1>
+            <h1 class="font-bold text-2xl mt-6 mb-1">Grupinė veikla "{{$groupActivity->title}}"</h1>
         @if($groupActivity->start_time==null || $currenttime < $groupActivity->start_time)
-            <h2>Talpa:{{$groupActivity->size}}</h2>
-            <h2>Liko vietų:{{$groupActivity->free_spaces}}</h2>
+            <h2>Talpa: {{$groupActivity->size}}</h2>
+            <h2>Liko vietų: {{$groupActivity->free_spaces}}</h2>
         @if($groupActivity->activityToGroup->notified == 1)
                 <div class="flex">
                 <p>Grupė jau&nbsp;</p>
                     <p class="text-green-700 font-bold">informuota&nbsp;</p>
-                <p>apie startą</p>
+                <p>apie startą.</p>
                 </div>
                 <h4>Starto data ir laikas: {{$groupActivity->start_time}}</h4>
             @else
             <div class="flex">
                 <p>Grupė&nbsp;</p>
                 <p class="text-red-700 font-bold">neinformuota&nbsp;</p>
-                <p>apie startą</p>
+                <p>apie startą.</p>
             </div>
             @endif
             @else
@@ -28,9 +28,9 @@
                     <p>galite išvalyti grupę.</p>
                 </div>
             @endif
-            <div class="overflow-x-auto relative border border-black">
+            <div class="overflow-x-auto relative border border-black rounded-lg mt-2">
                 <table class="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-                    <thead class="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400 border-b border-black">
+                    <thead class="text-xs text-gray-700 uppercase bg-gray-300 dark:bg-gray-700 dark:text-gray-400 border-b border-black">
                     <tr>
                         <th scope="col" class="py-3 px-6">
                             Vardas
@@ -58,7 +58,7 @@
                     <tbody>
                     @if($groupActivity->activityToGroup->groupToMember->count()>0)
                     @foreach($groupActivity->activityToGroup->groupToMember as $groupmember)
-                        <tr class="bg-gray-200 border-b dark:border-gray-700 text-gray-600">
+                        <tr class="bg-gray-100 border-b border-gray-300 dark:border-gray-700 text-gray-600">
                             <td
                                 class="py-4 px-6">
                                 {{$groupmember->name}}
@@ -84,7 +84,7 @@
                                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z"></path></svg>
                                 </button>
                                 </a>
-                                <a href="/deleteMember/{{$groupmember->id}}">
+                                <a onclick="return confirm('Ar tikrai norite pašalinti narį?')" href="/deleteMember/{{$groupmember->id}}">
                                 <button type="button" class="text-white bg-red-700 hover:bg-red-800 focus:outline-none focus:ring-4 focus:ring-red-300 font-medium rounded-full text-sm px-1 py-1 text-center mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">
                                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"></path></svg>
                                 </button>
@@ -93,10 +93,10 @@
                         </tr>
                     @endforeach
                     @else
-                        <tr class="bg-gray-200 border-b dark:border-gray-700">
+                        <tr class="bg-gray-100 dark:border-gray-700">
                             <th scope="row"
                                 class="py-4 px-6 font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                                Nėra užsiregistravusių narių
+                                Nėra užsiregistravusių narių.
                             </th>
                             <th>
                             </th>
